chore(server): migrate to FastAPI lifespan + harden /runs against path traversal
